Moved Webpack configs to /configs
