+++ configs/webpack.config.renderer.dev.dll.js
@@ -1,58 +1,32 @@
-/* eslint global-require: 0, import/no-dynamic-require: 0 */
-
 /**
- * Build config for development electron renderer process that uses
- * Hot-Module-Replacement
- *
- * https://webpack.js.org/concepts/hot-module-replacement/
+ * Builds the DLL for development electron renderer process
  */
 
-import path from 'path';
-import fs from 'fs';
 import webpack from 'webpack';
-import chalk from 'chalk';
+import path from 'path';
 import merge from 'webpack-merge';
-import { spawn, execSync } from 'child_process';
-import ExtractTextPlugin from 'extract-text-webpack-plugin';
 import baseConfig from './webpack.config.base';
-import CheckNodeEnv from './internals/scripts/CheckNodeEnv';
-import { paths } from './app/data/paths';
-import OMSFile from './app/data/utils/OMSFile';
+import { dependencies } from './package.json';
+import CheckNodeEnv from '../internals/scripts/CheckNodeEnv';
 
 CheckNodeEnv('development');
 
-const port = process.env.PORT || 1212;
-const publicPath = `http://localhost:${port}/dist`;
-const dll = path.resolve(process.cwd(), 'dll');
-const manifest = path.resolve(dll, 'renderer.json');
-
-let themePath = paths.appData + paths.theme;
-
-/**
- * Warn if the DLL is not built
- */
-if (!(fs.existsSync(dll) && fs.existsSync(manifest))) {
-    console.log(chalk.black.bgYellow.bold('The DLL files are missing. Sit back while we build them for you with "npm run build-dll"'));
-    execSync('npm run build-dll');
-}
+const dist = path.resolve(process.cwd(), 'dll');
 
 export default merge.smart(baseConfig, {
-    devtool: 'inline-source-map',
+    mode: 'development',
+
+    context: process.cwd(),
+
+    devtool: 'eval',
 
     target: 'electron-renderer',
 
-    entry: [
-        'react-hot-loader/patch',
-        `webpack-dev-server/client?http://localhost:${port}/`,
-        'webpack/hot/only-dev-server',
-        path.join(__dirname, 'app/index.js')
-    ],
+    externals: ['fsevents', 'crypto-browserify'],
 
-    output: {
-        publicPath: `http://localhost:${port}/dist/`,
-        filename: 'renderer.dev.js'
-    },
-
+    /**
+     * Use `module` from `webpack.config.renderer.dev.js`
+     */
     module: {
         rules: [
             {
@@ -69,7 +43,7 @@ export default merge.smart(baseConfig, {
                             'transform-class-properties',
                             'transform-es2015-classes',
                             'react-hot-loader/babel'
-                        ]
+                        ],
                     }
                 }
             },
@@ -125,9 +99,6 @@ export default merge.smart(baseConfig, {
                                 path.resolve(__dirname, "node_modules")
                             ],
                             javascriptEnabled: true,
-                            modifyVars: {
-                                '@theme': OMSFile.readJSON(themePath).theme
-                            }
                         }
                     }
                 ]
@@ -139,9 +110,9 @@ export default merge.smart(baseConfig, {
                     loader: 'url-loader',
                     options: {
                         limit: 10000,
-                        mimetype: 'application/font-woff'
+                        mimetype: 'application/font-woff',
                     }
-                }
+                },
             },
             // WOFF2 Font
             {
@@ -150,7 +121,7 @@ export default merge.smart(baseConfig, {
                     loader: 'url-loader',
                     options: {
                         limit: 10000,
-                        mimetype: 'application/font-woff'
+                        mimetype: 'application/font-woff',
                     }
                 }
             },
@@ -168,101 +139,69 @@ export default merge.smart(baseConfig, {
             // EOT Font
             {
                 test: /\.eot(\?v=\d+\.\d+\.\d+)?$/,
-                use: 'file-loader'
+                use: 'file-loader',
             },
             // SVG Font
             {
                 test: /\.svg(\?v=\d+\.\d+\.\d+)?$/,
                 use: {
-                    loader: 'url-loader',
-                    options: {
-                        limit: 10000,
-                        mimetype: 'image/svg+xml'
-                    }
+                loader: 'url-loader',
+                options: {
+                    limit: 10000,
+                    mimetype: 'image/svg+xml',
+                }
                 }
             },
             // Common Image Formats
             {
                 test: /\.(?:ico|gif|png|jpg|jpeg|webp)$/,
-                use: 'url-loader'
+                use: 'url-loader',
             }
         ]
     },
 
+    entry: {
+        renderer: (
+            Object
+            .keys(dependencies || {})
+            .filter(dependency => dependency !== 'font-awesome')
+        )
+    },
+
+    output: {
+        library: 'renderer',
+        path: dist,
+        filename: '[name].dev.dll.js',
+        libraryTarget: 'var'
+    },
+
     plugins: [
-    new webpack.DllReferencePlugin({
-        context: process.cwd(),
-        manifest: require(manifest),
-        sourceType: 'var'
-    }),
+        new webpack.DllPlugin({
+            path: path.join(dist, '[name].json'),
+            name: '[name]',
+        }),
 
-    new webpack.HotModuleReplacementPlugin({
-        multiStep: true
-    }),
+        /**
+         * Create global constants which can be configured at compile time.
+         *
+         * Useful for allowing different behaviour between development builds and
+         * release builds
+         *
+         * NODE_ENV should be production so that modules do not perform certain
+         * development checks
+         */
+        new webpack.EnvironmentPlugin({
+            NODE_ENV: 'development'
+        }),
 
-    new webpack.NoEmitOnErrorsPlugin(),
-
-    /**
-     * Create global constants which can be configured at compile time.
-     *
-     * Useful for allowing different behaviour between development builds and
-     * release builds
-     *
-     * NODE_ENV should be production so that modules do not perform certain
-     * development checks
-     *
-     * By default, use 'development' as NODE_ENV. This can be overriden with
-     * 'staging', for example, by changing the ENV variables in the npm scripts
-     */
-    new webpack.EnvironmentPlugin({
-        NODE_ENV: 'development'
-    }),
-
-    new webpack.LoaderOptionsPlugin({
-        debug: true
-    }),
-
-    new ExtractTextPlugin({
-        filename: '[name].css'
-    }),
-    ],
-
-    node: {
-        __dirname: false,
-        __filename: false
-    },
-
-    devServer: {
-        port,
-        publicPath,
-        compress: true,
-        noInfo: true,
-        stats: 'errors-only',
-        inline: true,
-        lazy: false,
-        hot: true,
-        headers: { 'Access-Control-Allow-Origin': '*' },
-        contentBase: path.join(__dirname, 'dist'),
-        watchOptions: {
-            aggregateTimeout: 300,
-            ignored: /node_modules/,
-            poll: 100
-        },
-        historyApiFallback: {
-            verbose: true,
-            disableDotRule: false,
-        },
-        before() {
-            if (process.env.START_HOT) {
-            console.log('Starting Main Process...');
-            spawn(
-                'npm',
-                ['run', 'start-main-dev'],
-                { shell: true, env: process.env, stdio: 'inherit' }
-            )
-                .on('close', code => process.exit(code))
-                .on('error', spawnError => console.error(spawnError));
+        new webpack.LoaderOptionsPlugin({
+            debug: true,
+            options: {
+                context: path.resolve(process.cwd(), 'app'),
+                output: {
+                    path: path.resolve(process.cwd(), 'dll'),
+                },
             }
-        }
-    },
+        })
+    ]
 });
